requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.1.0
torch-geometric>=2.4.0
lifelines>=0.27.0
scikit-survival>=0.22.0
matplotlib>=3.7.0
seaborn>=0.12.0
shap>=0.42.0
pyyaml
tqdm
joblib

# Multi-Agent System Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
